Implemented Overworld and Chunk class. Structure is now as follows: Canvas: 	-Chunk: 		-Overworld 		-Dungeon Array
However, this produces a seg fault, which seems to originate in the dungeon generation,
but compiles fine.  It has something to do with getting out of the range of the vectors,
but I was unable to see exactly where.

There is also a memory leak which sometimes occurs, seemingly at random.

Recommended testing with Valgrind.