Beating my head against the wall with resizing in the x direction for chunks.
Characters can move freely through chunks (down to an index of 0) in the y
direction, but as soon as they go out of the initially declared size of 10
in the x direction, the characters current chunk goes out of range.  I have
no idea where this could be happening...ugg.