Fixed a bug in the enemy so that the they can no longer move through trees.
There is an inconsistant segfault caused by get_surroundings, where an enemy is
too close to the edge of a chunk which is either unitialized or out of range of the
chunk_map.  This is obviously a problem, and the solution that I came up with was to just
have that enemy update the chunk_map.  This created a whole slew of memory errors and an
even worse bug, so I reverted it.  I'll keep thinking of a better way to do it.