Enemy speed is now regulated by actual time passage. Enemies have a speed attribute, which can be interpreted as 'milliseconds per step'. Higher is slower. Some thoughts:
We have to think about how to make this a frame-slowing instead of
frame-skipping application. In other words, when the user's computer
cannot keep up with the intended speed (which is what? How do we define
that?), it is important that we slow down the game instead of skipping
frames (frame skipping is what you would see in most modern games). This
is important because the combat will likely be too difficult if things
are jumping everywhere.

It doesn't seem likely at this point that a user's computer would have
trouble keeping up with what's going on (my laptop can handle it with a
very large margin), but who knows what the future may hold?