Reworked the debug console.  Now, you just have to add things into the debug.cpp file.  The way you have to do it is a little bit gross, but the code looks much better than when it was just a massive dump.
Basically, if you want to add a message, add it to db_messages (making
sure to change the size of the array) and then add it to MESSAGES in
debug.h.  If you want to add a function, add the function and then add
it to the mapping.  It's that simple.  I just need to remember to make
a way to scroll through pages of text.