Some more stuff with the buildings...there's an algorithm I need to generate houses, but I can't find it.  I'll probably have to make one up myself, but that means it'll take me a while.  I also made the window resizable, which I then realized could be used for the UI.  It makes it somewhat restrictive, as you need a minimum screen size to play anything (it wouldn't be playable on anything less than 600x1000 or so.  More thought, etc.
Basically, we could have our game running in 600x800, and then have 200 or so
(which would correspond to 25 characters) px designated for the UI.  This could
then be resized by the user to fit their screen.  We could stick in a screen resize
event, which would change our SCREEN_SIZE.  We could then use that for text
wrapping for the UI.  Or something along those lines.  Then people could make
the screen as large as they wanted, but there would be a minimun size for it
as well.