Started working on the new ai system.  I like the way it's shaping up, it's looking like it will be suuuuper extendable!  However, Xmax is on the way and I don't know how much more I'm going to get done before then. Also, the enemy class is a mess and this will NOT compile.
The new AI system will be using behavior trees and, for now, just the basic
set of nodes.  Essentially, all the behavior tree is for is making decisions.
It will NOT actually alter anything or assign any values--that is for methods
in the Game class (which also needs cleaned up).  Each node, upon acting, will
take in an actor to act upon which will be passed into an appropriate method in
the Game class.